+++ CHANGES.md
@@ -2,3 +2,4 @@ This build requires **[IP v5.2.5](https://github.com/Efnilite/Walk-in-the-Park/r
 
 - Fixed speed effect not being reset when switching mode
 - Fixed players dying when game of Duels ends 
+- Fixed invite message not working on Spigot